refactor(istream): tighten sofab_istream context member types & packing
Audit the sofab_istream fields against the config options that govern their
value ranges, apply the matching typedefs, and pack the struct by alignment so
its footprint stays minimal — most notably on the 16-bit-size_t MCU targets the
library is designed for.

- id: uint32_t -> sofab_id_t. sofab_id_t is the config-owned field-id type
  (SOFAB_ID_MAX); the decoder already cast to it and the field callback already
  takes it, so the member now tracks that type instead of a hardcoded width.

- fixlen_remaining / target_len / target_count / array_wire_count:
  uint32_t -> size_t. These hold byte lengths and element counts bounded by
  SOFAB_FIXLEN_MAX / SOFAB_ARRAY_MAX (both derived from SIZE_MAX) and are fed
  straight from the size_t read-function arguments, so they now mirror that
  width exactly. This drops the three truncating (uint32_t) casts on the array
  paths (also cutting the -Wconversion count 7 -> 2) and narrows the fields on
  16-bit-size_t targets (uint32_t there was wider than the values can ever be).

- Regroup members by alignment (value-word / pointers / size_t, then the 32-bit
  id, then the 8-bit flags) so the opaque context packs with no internal
  padding. On 64-bit this closes the 4-byte hole before target_ptr; the struct
  is documented opaque, so field order carries no meaning.

Net footprint: ~8 bytes smaller on 16-bit MCUs (AVR/RL78), unchanged on 32-bit
MCUs, +8 bytes on 64-bit hosts (the size_t widening, accepted since hosts are
not footprint-bound). Layout/ABI shifts with config width, consistent with the
library's existing "configs are not ABI-compatible" stance.

All C/C++ test suites pass; istream.c builds warning-free under
-Wall -Wextra -Wpedantic -Werror across every config toggle.

+++ src/istream.c
@@ -344,7 +344,7 @@ static sofab_ret_t _bind_array_count (sofab_istream_t *ctx, size_t wire_count)
         }
 
         // read exactly what the wire carries; trailing slots keep their defaults
-        ctx->target_count = (uint32_t)wire_count;
+        ctx->target_count = wire_count;
     }
 
     return SOFAB_RET_OK;
@@ -828,8 +828,8 @@ extern sofab_ret_t sofab_istream_feed (sofab_istream_t *ctx, const void *data, s
                 // and, for fixlen arrays, the hop to _DECODER_STATE_FIXLEN_LEN
                 // where the destination is actually bound.
                 size_t count = (size_t)(array_count);
-                ctx->target_count = (uint32_t)count;
-                ctx->array_wire_count = (uint32_t)count;
+                ctx->target_count = count;
+                ctx->array_wire_count = count;
 
                 // read array elements (see IDLE state for type check)
                 uint8_t type = _OPT_FIELDTYPE(ctx->target_opt);
